+++ packages/analyzer_kit_annotation/lib/analyzer_kit_annotation.dart
@@ -6,5 +6,7 @@ library;
 export 'src/analyzer_kit.dart';
 export 'src/copy_with.dart';
 export 'src/data_class.dart';
+export 'src/deserialize.dart';
 export 'src/override_equality.dart';
 export 'src/override_to_string.dart';
+export 'src/serialize.dart';