+++ packages/analyzer_kit_annotation/CHANGELOG.md
@@ -5,3 +5,5 @@
 - Added `OverrideEquality` annotation for `==` and `hashCode` generation.
 - Added `OverrideToString` annotation for `toString` generation.
 - Added `DataClass` annotation as a shorthand to generate `copyWith`, `==`, `hashCode`, and `toString`.
+- Added `Serialize` annotation for map extraction generation.
+- Added `Deserialize` annotation for factory map constructor generation.